Fix my Entries index turbo_frame implementation
I was going about this a bit wrong and it was resulting in goofy
rendering problems when creating new Entries. A little google-fu and a
StackOverflow got me going in the right direction.

Key points:
`target: "_top"` - because you want other links to navigate outside of a
frame and pagination links to be within a frame.
`turbo_action: "advance"` - updates current url.